Circuits for the SMRT-Y snitch
First simple version had trouble maintaining the virtual ground.
It was likely a combination of three things: a wiring error that
connected somethings to the "real" ground not the "virtual ground",
overly-large resistors in the reference voltage divider (TL084 has
input impedance which is significant in this case), and the TL084
misbehaving when run as a comparator (it doesn't like its outputs
to be saturated).

SMRTYSnitch2 used a fancier virtual ground circuit as step 1 in
addressing this, and reduced the voltage reference voltage divider
to 100k on each side (from 1M).

SMRTYSnith3 switched to using a 'real' comparator (LM339 is what
I had on hand), which addressed the problems 'properly'.